web: add disabled tests
The idea is to fix them ASAP
